Fix barcode scanner preview orientation on Nexus 5x
The code that is supposed to deal with sensor orientation has a bug. The findCamera() method
never calls Camera#getCameraInfo() method for the first camera (index 0). As a result, mCameraInfo
may never be set to reflect the properties of the camera identified with mCameraId (may instead
contain info of the other camera).